+++ domain/src/main/kotlin/com/sooum/domain/usecase/auth/LoginUseCase.kt
@@ -1,4 +1,4 @@
-package com.sooum.domain.usecase.meet
+package com.sooum.domain.usecase.auth
 
 import com.sooum.domain.model.ApiResult
 import com.sooum.domain.repository.AuthRepository
